service: Promote Bootloader::Config
