+++ service/lib/agama/storage/bootloader.rb
@@ -20,11 +20,11 @@
 # find current contact information at www.suse.com.
 
 require "yast"
-require "json"
 require "bootloader/bootloader_factory"
 require "bootloader/os_prober"
 
 require "agama/http/clients"
+require "agama/storage/bootloader_config"
 
 Yast.import "BootStorage"
 
@@ -32,104 +32,12 @@ module Agama
   module Storage
     # Represents bootloader specific functionality
     class Bootloader
-      # Represents bootloader settings
-      class Config
-        # Whether bootloader should update persistent RAM.
-        # If kept as nil then default will be used.
-        #
-        # @return [Boolean]
-        attr_accessor :update_nvram
-
-        # Whether bootloader should stop on boot menu.
-        #
-        # @return [Boolean]
-        attr_accessor :stop_on_boot_menu
-
-        # Bootloader timeout.
-        #
-        # Only positive numbers are supported and stop_on_boot_menu has precedence.
-        #
-        # @return [Integer]
-        attr_accessor :timeout
-
-        # Bootloader extra kernel parameters beside ones that is proposed.
-        #
-        # @return [String]
-        attr_accessor :extra_kernel_params
-
-        # Bootloader extra kernel parameters needed for other parts of agama
-        #
-        # @return [Hash<String, String>]
-        attr_accessor :scoped_kernel_params
-
-        # Keys to export to JSON.
-        #
-        # As both previous keys are conflicting, remember which one to set or none. It can be empty
-        # and it means export nothing.
-        #
-        # @return [Array<Symbol>]
-        attr_accessor :keys_to_export
-
-        def initialize
-          @keys_to_export = []
-          @stop_on_boot_menu = false # false means use proposal, which has timeout
-          @timeout = 10 # just some reasonable timeout, we do not send it anywhere
-          @update_nvram = nil
-          @extra_kernel_params = ""
-          @scoped_kernel_params = {}
-        end
-
-        # Serializes the config to JSON.
-        #
-        # @return [String]
-        def to_json(*_args)
-          result = {}
-
-          # our json use camel case
-          result[:stopOnBootMenu] = stop_on_boot_menu if keys_to_export.include?(:stop_on_boot_menu)
-          result[:timeout] = timeout if keys_to_export.include?(:timeout)
-          result[:updateNvram] = update_nvram if keys_to_export.include?(:update_nvram)
-          if keys_to_export.include?(:extra_kernel_params)
-            result[:extraKernelParams] =
-              @extra_kernel_params
-          end
-
-          result.to_json
-        end
-
-        # Loads the config from a JSON string.
-        #
-        # @param serialized_config [String]
-        # @return [Config] self
-        def load_json(serialized_config)
-          hsh = JSON.parse(serialized_config, symbolize_names: true)
-          update_attribute(hsh, :timeout, :timeout, conflicts: :stop_on_boot_menu)
-          update_attribute(hsh, :stopOnBootMenu, :stop_on_boot_menu, conflicts: :timeout)
-          update_attribute(hsh, :extraKernelParams, :extra_kernel_params)
-          update_attribute(hsh, :updateNvram, :update_nvram)
-
-          self.scoped_kernel_params = hsh[:kernelArgs]
-
-          self
-        end
-
-      private
-
-        def update_attribute(hsh, json_key, attr_key, conflicts: nil)
-          return unless hsh.key?(json_key)
-
-          public_send("#{attr_key}=", hsh[json_key])
-          keys_to_export.delete(conflicts) if conflicts
-          keys_to_export.push(attr_key) unless keys_to_export.include?(attr_key)
-        end
-      end
-
-      # @return [Config]
+      # @return [BootloaderConfig]
       attr_reader :config
 
       # @param logger [Logger]
       def initialize(logger)
-        @config = Config.new
+        @config = BootloaderConfig.new
         @logger = logger
       end
