+++ service/test/agama/storage/bootloader_test.rb
@@ -25,49 +25,6 @@ require "agama/storage/bootloader"
 require "bootloader/grub2"
 require "bootloader/systemdboot"
 
-describe Agama::Storage::Bootloader::Config do
-  subject(:config) { described_class.new }
-
-  describe "#to_json" do
-    before do
-      config.load_json({ "stopOnBootMenu" => true }.to_json)
-    end
-
-    it "serializes its content with keys as camelCase" do
-      expect(config.to_json).to eq "{\"stopOnBootMenu\":true}"
-    end
-
-    it "can serialize in a way that #load_json can restore it" do
-      config.stop_on_boot_menu = false
-      json = config.to_json
-      config.stop_on_boot_menu = true
-      config.load_json(json)
-      expect(config.stop_on_boot_menu).to eq false
-    end
-
-    it "exports only what was previously set" do
-      expect(config.to_json).to eq "{\"stopOnBootMenu\":true}"
-      config.load_json({ "timeout" => 10, "extraKernelParams" => "verbose" }.to_json)
-      expect(config.to_json).to eq "{\"timeout\":10,\"extraKernelParams\":\"verbose\"}"
-    end
-  end
-
-  describe "#load_json" do
-    it "loads config from given json" do
-      content = "{\"stopOnBootMenu\":true,\"updateNvram\":true}"
-      config.load_json(content)
-      expect(config.stop_on_boot_menu).to eq true
-      expect(config.update_nvram).to eq true
-    end
-
-    it "remembers which keys are set" do
-      content = "{\"timeout\":10}"
-      config.load_json(content)
-      expect(config.keys_to_export).to eq([:timeout])
-    end
-  end
-end
-
 describe Agama::Storage::Bootloader do
   let(:logger) { Logger.new($stdout, level: :warn) }
   let(:agama_bootloader) { described_class.new(logger) }
